pyproject.toml
[project]
name = "forge"
version = "0.1.0"
description = "Forge — autonomous lab assignment runtime over Daytona sandboxes."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "daytona>=0.167.0",
    "pyyaml>=6.0.2",
]

[project.scripts]
forge-runtime = "forge_runtime.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["forge_runtime"]
